Переносит функцию в файл и подключает в точку входа
